refactor: modify type and data and filename and func name
