+++ frontend/src/util/toArray.ts
@@ -1,3 +1,3 @@
-export function isArray<T>(input: T | T[]): T[] {
+export function toArray<T>(input: T | T[]): T[] {
   return Array.isArray(input) ? input : [input];
 }
